Move into functions into sub module
Create an into module and move the into functions there. Code move only.
